add bar background clipping
Implements the features discussed in #274. It allows the clipping of the
bar background to reveal what is underneath. The clipping shape is based
on the background dimensions & corner radius, and the clipping mask is a
float (0.f = no clipping, 1.f = fully clipped). The background.clip
property is fully animatable, and can be applied to items and brackets,
but not to the bar itself or to text items.

Floating point number properties can now be animated using the
ANIMATE_FLOAT macro.

Clipping should not affect performance when disabled, aside from an
additional loop in the bar_draw function. When enabled, there are at the
moment quite a few preventable bar redraws since the bar_needs_refresh
is naive and does not check if the bar actually needs a refresh (i.e.
the clipping value has changed, the item has moved or its dimensions
were changed). A less aggressive change detection may be implemented but
would require a bit of refactoring to the update functions.