only try locking the mutex for an animation frame
the animation_frame_callback function shall never be blocked by waiting for the mutex to unlock.
This is because the CVDiplayLink* funtions will try to acquire a mutex that is locked during the
call to animation_frame_callback. This can lead to a dead lock scenario when one thread handles an
event, the animation thread waits for this event to be handled but the handling of said event requires
to unlock the CVDisplayLink mutex...