reimplement the sticky property
The new sticky implementation makes the bar fully sticky, even if space change animations are enabled, it will not move together with the spaces anymore.

The sticky property was forcibly enabled on Sonoma up till this point, now the sticky property can be disabled again, but it defaults to being on.